+++ dflib/src/main/java/com/nhl/dflib/LongSeries.java
@@ -2,6 +2,7 @@ package com.nhl.dflib;
 
 import com.nhl.dflib.builder.BoolAccum;
 import com.nhl.dflib.builder.LongAccum;
+import com.nhl.dflib.series.LongArraySeries;
 
 import java.util.Comparator;
 import java.util.Random;
@@ -51,6 +52,30 @@ public interface LongSeries extends Series<Long> {
     void copyToLong(long[] to, int fromOffset, int toOffset, int len);
 
     LongSeries materializeLong();
+
+    /**
+     * @since 0.18
+     */
+    @Override
+    default Series<?> add(Object value) {
+        return value instanceof Long
+                ? addLong((Long) value)
+                : Series.super.add(value);
+    }
+
+    /**
+     * Creates a new Series with a provided int appended to the end of this Series.
+     *
+     * @since 0.18
+     */
+    default LongSeries addLong(long val) {
+        int s = size();
+
+        long[] data = new long[s + 1];
+        this.copyToLong(data, 0, 0, s);
+        data[s] = val;
+        return new LongArraySeries(data);
+    }
 
     LongSeries concatLong(LongSeries... other);